Migrate the navigation logic in Home screen from androidx.navigation to voyager
