+++ app/src/main/java/com/mr3y/ludi/ui/navigation/PreferencesType.kt
@@ -1,0 +1,7 @@
+package com.mr3y.ludi.ui.navigation
+
+enum class PreferencesType {
+    NewsDataSources,
+    Genres,
+    Games
+}
